# JoinOrderingRule