--- a/src/lib/cost_model/abstract_cost_estimator.cpp
+++ b/src/lib/cost_model/abstract_cost_estimator.cpp
@@ -8,8 +8,10 @@
 namespace opossum {
 
 Cost AbstractCostEstimator::estimate_plan_cost(const std::shared_ptr<AbstractLQPNode>& lqp) const {
-  // Cost all operators in arbitrary order
+  // Sum up the costs of all operators in the plan, while making sure to cost each operator exactly once, even in the
+  // presence of diamond shapes.
 
+  // Breadth-first iteration of plan
   auto bfs_queue = std::queue<std::shared_ptr<AbstractLQPNode>>{};
   auto visited = std::unordered_set<std::shared_ptr<AbstractLQPNode>>{};