--- a/src/lib/cost_model/abstract_cost_model.cpp
+++ b/src/lib/cost_model/abstract_cost_model.cpp
@@ -1,65 +1,33 @@
 #include "abstract_cost_model.hpp"
 
-#include "cost_feature_lqp_node_proxy.hpp"
-#include "cost_feature_operator_proxy.hpp"
+#include <unordered_set>
+#include <queue>
+
 #include "logical_query_plan/abstract_lqp_node.hpp"
-#include "logical_query_plan/join_node.hpp"
-#include "logical_query_plan/predicate_node.hpp"
-#include "logical_query_plan/union_node.hpp"
-#include "operators/join_hash.hpp"
-#include "operators/operator_join_predicate.hpp"
-#include "operators/operator_scan_predicate.hpp"
-#include "operators/product.hpp"
-#include "operators/table_scan.hpp"
-#include "operators/union_positions.hpp"
 
 namespace opossum {
 
-Cost AbstractCostModel::estimate_operator_cost(const std::shared_ptr<AbstractOperator>& op) const {
-  CostFeatureOperatorProxy feature_proxy(op);
-  return _cost_model_impl(op->type(), feature_proxy);
-}
+Cost AbstractCostModel::estimate_plan_cost(const std::shared_ptr<AbstractLQPNode>& lqp) const {
+  // Cost all operators in arbitrary order
 
-Cost AbstractCostModel::estimate_lqp_node_cost(const std::shared_ptr<AbstractLQPNode>& node) const {
-  auto operator_type = OperatorType::Mock;  // Use Mock just to have a valid initialisation
+  auto bfs_queue = std::queue<std::shared_ptr<AbstractLQPNode>>{};
+  auto visited = std::unordered_set<std::shared_ptr<AbstractLQPNode>>{};
 
-  /**
-   * The following switch makes assumptions about the concrete Operator that the LQPTranslator will choose.
-   * TODO(anybody) somehow ask the LQPTranslator about this instead of making assumptions.
-   */
+  bfs_queue.emplace(lqp);
 
-  switch (node->type) {
-    case LQPNodeType::Predicate:
-      operator_type = OperatorType::TableScan;
-      break;
+  auto cost = Cost{0};
 
-    case LQPNodeType::Join: {
-      const auto join_node = std::static_pointer_cast<JoinNode>(node);
-      const auto operator_predicate = OperatorJoinPredicate::from_expression(
-          *join_node->join_predicate, *join_node->left_input(), *join_node->right_input());
-      Assert(operator_predicate, "Expected Join predicate to be OperatorScanPredicate compatible");
+  while (!bfs_queue.empty()) {
+    const auto current_node = bfs_queue.front();
+    bfs_queue.pop();
+    if (!visited.emplace(current_node).second) {
+      continue;
+    }
 
-      if (join_node->join_mode == JoinMode::Cross) {
-        operator_type = OperatorType::Product;
-      } else if (join_node->join_mode == JoinMode::Inner &&
-                 operator_predicate->predicate_condition == PredicateCondition::Equals) {
-        operator_type = OperatorType::JoinHash;
-      } else {
-        operator_type = OperatorType::JoinSortMerge;
-      }
-    } break;
-
-    case LQPNodeType::Union: {
-      operator_type = OperatorType::UnionPositions;
-    } break;
-
-    default:
-      // TODO(anybody) we're not costing this OperatorType yet (since it is not involved in JoinOrdering and all
-      //               costing is currently done for JoinOrdering only)
-      return 0.0f;
+    cost += _estimate_node_cost(current_node);
   }
 
-  CostFeatureLQPNodeProxy feature_proxy(node);
-  return _cost_model_impl(operator_type, feature_proxy);
+  return cost;
 }
+
 }  // namespace opossum