--- a/src/lib/cost_model/cost_model_logical.cpp
+++ b/src/lib/cost_model/cost_model_logical.cpp
@@ -1,44 +1,36 @@
 #include "cost_model_logical.hpp"
 
-#include "abstract_cost_feature_proxy.hpp"
-#include "operators/abstract_operator.hpp"
+#include "logical_query_plan/abstract_lqp_node.hpp"
+#include "logical_query_plan/union_node.hpp"
+#include "logical_query_plan/join_node.hpp"
+#include "statistics/table_statistics.hpp"
+#include "utils/assert.hpp"
 
 namespace opossum {
 
-std::string CostModelLogical::name() const { return "CostModelLogical"; }
+Cost CostModelLogical::_estimate_node_cost(const std::shared_ptr<AbstractLQPNode>& node) const {
+  const auto output_row_count = node->get_statistics()->row_count();
+  const auto left_row_count = node->left_input()->get_statistics()->row_count();
+  const auto right_row_count = node->right_input() ? node->right_input()->get_statistics()->row_count() : 0.0f;
 
-Cost CostModelLogical::get_reference_operator_cost(const std::shared_ptr<AbstractOperator>& op) const {
-  return estimate_operator_cost(op);
-}
+  switch (node->type) {
+    case LQPNodeType::Join:
+      return left_row_count + right_row_count + output_row_count;
 
-Cost CostModelLogical::_cost_model_impl(const OperatorType operator_type,
-                                        const AbstractCostFeatureProxy& feature_proxy) const {
-  switch (operator_type) {
-    case OperatorType::JoinHash:
-      return feature_proxy.extract_feature(CostFeature::LeftInputRowCount).scalar() +
-             feature_proxy.extract_feature(CostFeature::RightInputRowCount).scalar();
+    case LQPNodeType::Sort:
+      return left_row_count * std::log(left_row_count);
 
-    case OperatorType::TableScan:
-      return feature_proxy.extract_feature(CostFeature::LeftInputRowCount).scalar();
+    case LQPNodeType::Union: {
+      const auto union_node = std::static_pointer_cast<UnionNode>(node);
 
-    case OperatorType::JoinSortMerge:
-      // Model the cost of the sorting as the dominant cost
-      return feature_proxy.extract_feature(CostFeature::LeftInputRowCountLogN).scalar() +
-             feature_proxy.extract_feature(CostFeature::RightInputRowCountLogN).scalar();
-
-    case OperatorType::Product:
-      return feature_proxy.extract_feature(CostFeature::InputRowCountProduct).scalar();
-
-    case OperatorType::JoinNestedLoop:
-      return feature_proxy.extract_feature(CostFeature::InputRowCountProduct).scalar();
-
-    case OperatorType::UnionPositions:
-      // Model the cost of the sorting as the dominant cost
-      return feature_proxy.extract_feature(CostFeature::LeftInputRowCountLogN).scalar() +
-             feature_proxy.extract_feature(CostFeature::RightInputRowCountLogN).scalar();
+      switch(union_node->union_mode) {
+        case UnionMode::Positions:
+          return left_row_count * std::log(left_row_count) + right_row_count * std::log(right_row_count);
+      }
+    }
 
     default:
-      return 0.0f;
+      return left_row_count + output_row_count;
   }
 }