--- a/src/lib/cost_model/cost_model_logical.cpp
+++ b/src/lib/cost_model/cost_model_logical.cpp
@@ -13,7 +13,7 @@ namespace opossum {
 
 Cost CostModelLogical::_estimate_node_cost(const std::shared_ptr<AbstractLQPNode>& node) const {
   const auto output_row_count = node->get_statistics()->row_count();
-  const auto left_row_count = node->left_input()->get_statistics()->row_count();
+  const auto left_row_count = node->left_input() ? node->left_input()->get_statistics()->row_count() : 0.0f;
   const auto right_row_count = node->right_input() ? node->right_input()->get_statistics()->row_count() : 0.0f;
 
   switch (node->type) {