--- a/src/lib/cost_model/cost_model_logical.cpp
+++ b/src/lib/cost_model/cost_model_logical.cpp
@@ -13,22 +13,24 @@ namespace opossum {
 
 Cost CostModelLogical::_estimate_node_cost(const std::shared_ptr<AbstractLQPNode>& node) const {
   const auto output_row_count = node->get_statistics()->row_count();
-  const auto left_row_count = node->left_input() ? node->left_input()->get_statistics()->row_count() : 0.0f;
-  const auto right_row_count = node->right_input() ? node->right_input()->get_statistics()->row_count() : 0.0f;
+  const auto left_input_row_count = node->left_input() ? node->left_input()->get_statistics()->row_count() : 0.0f;
+  const auto right_input_row_count = node->right_input() ? node->right_input()->get_statistics()->row_count() : 0.0f;
 
   switch (node->type) {
     case LQPNodeType::Join:
-      return left_row_count + right_row_count + output_row_count;
+      // Covers predicated and unpredicated joins. For cross joins, output_row_count will be
+      // left_input_row_count * right_input_row_count
+      return left_input_row_count + right_input_row_count + output_row_count;
 
     case LQPNodeType::Sort:
-      return left_row_count * std::log(left_row_count);
+      return left_input_row_count * std::log(left_input_row_count);
 
     case LQPNodeType::Union: {
       const auto union_node = std::static_pointer_cast<UnionNode>(node);
 
       switch (union_node->union_mode) {
         case UnionMode::Positions:
-          return left_row_count * std::log(left_row_count) + right_row_count * std::log(right_row_count);
+          return left_input_row_count * std::log(left_input_row_count) + right_input_row_count * std::log(right_input_row_count);
         default:
           Fail("GCC thinks this is reachable");
       }
@@ -36,15 +38,17 @@ Cost CostModelLogical::_estimate_node_cost(const std::shared_ptr<AbstractLQPNode
 
     case LQPNodeType::Predicate: {
       const auto predicate_node = std::static_pointer_cast<PredicateNode>(node);
-      return left_row_count * _get_expression_cost_multiplier(predicate_node->predicate) + output_row_count;
+      return left_input_row_count * _get_expression_cost_multiplier(predicate_node->predicate) + output_row_count;
     }
 
     default:
-      return left_row_count + output_row_count;
+      return left_input_row_count + output_row_count;
   }
 }
 
 float CostModelLogical::_get_expression_cost_multiplier(const std::shared_ptr<AbstractExpression>& expression) {
+  // Number of operations +  number of different columns accessed to factor in expression complexity
+
   auto multiplier = 0.0f;
 
   visit_expression(expression, [&](const auto& sub_expression) {