--- a/src/lib/cost_model/cost_model_logical.hpp
+++ b/src/lib/cost_model/cost_model_logical.hpp
@@ -7,13 +7,14 @@ namespace opossum {
 class AbstractExpression;
 
 /**
- * Logical complexity cost model
+ * Cost model for logical complexity, i.e., approximate number of tuple accesses
  */
 class CostModelLogical : public AbstractCostModel {
  protected:
   Cost _estimate_node_cost(const std::shared_ptr<AbstractLQPNode>& node) const override;
 
-  // Number of operations +  number of different columns accessed to judge expression complexity
+ private:
+  // Number of operations +  number of different columns accessed to factor in expression complexity
   static float _get_expression_cost_multiplier(const std::shared_ptr<AbstractExpression>& expression);
 };