--- a/src/lib/cost_model/cost_model_logical.hpp
+++ b/src/lib/cost_model/cost_model_logical.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "abstract_cost_model.hpp"
+#include "abstract_cost_estimator.hpp"
 
 namespace opossum {
 
@@ -9,7 +9,7 @@ class AbstractExpression;
 /**
  * Cost model for logical complexity, i.e., approximate number of tuple accesses
  */
-class CostModelLogical : public AbstractCostModel {
+class CostModelLogical : public AbstractCostEstimator {
  protected:
   Cost _estimate_node_cost(const std::shared_ptr<AbstractLQPNode>& node) const override;