@mrzzzrm 2018-09-03
Two main issues raised here:

### Documentation of `DpCcp` and `EnumerateCcp`
I added some more comments, more out of goodwill than out of belief that this will actually help understanding the algorithm. 
I spend, I think, 10 pages of my MA explaining this algorithm. I cannot hope to boil that down to something that I can put in here as comment without distorting its content.

My stance here is the following: 
* Someone who has read and understood the paper will have no problem following this implementation since all concepts used directly map to concepts in the paper. Any comment explaining DPccp would just be noise for this person.
* Someone who has not understood the algorithm, frankly, doesn't stand a chance understanding it from any commenting I might add. I cannot magically make understanding of a subtle graph-processing algorithm appear in some readers brain just from a few comments, I am sorry. 

I recognize that this makes `DpCcp` and `EnumerateCcp` hard to review in depth, and I cannot offer a satisfying solution for this. Some things are just not simple, and I am afraid this is one of them.

### Testing of `DpCcp` and the `JoinOrderingRule`

@SvenLehmann rightly pointed out that there are not that many tests for `DpCcp` and `JoinOrderingRule`. I will add one or two more tests for `DpCcp` but stand by my point that it is not a good idea to test these too heavily since they heavily rely on the behaviour of the underlying cost model and cardinality estimators and therefore there is no clear definition of what a correctly functioning `DpCcp` or `JoinOrderingRule` entails. Any test I implement will likely need to be adjusted every time these estimators are changed. 
The bulk of correctness for `DpCcp` and therefore the `JoinOrderingRule` is ensured by the `EnumerateCcp`-Test.