@mrks 2018-09-04
I get your point about the documentation. Maybe it would help to have a high-level explanation of what the different classes do. For example, would it be possible to add a couple of lines to `enumerate_ccp.hpp` explaining what the algorithm takes as input and what its output is? It should at least be possible to understand the public interfaces of the different classes and what they do. *How* they do it is something that people can look up in the paper.