@mrzzzrm 2018-09-04
More tests added, I think that's sufficient now. Also added comments explaining why there are so "few" tests for the JoinOrderingRule and DpCcp